requirements.txt
# pandas is a data manipulation library
pandas==2.1.0

# numpy is a numerical computing library
numpy==1.24.3

# psycopg2-binary is a PostgreSQL integration library
psycopg2-binary==2.9.9

# SQLAlchemy is a SQL library
SQLAlchemy==2.0.21

# GeoAlchemy2 is a spatial library(geographic data)
GeoAlchemy2==0.14.1

# python-dotenv is a library for environment variables
python-dotenv==1.0.0

# requests is a library for API requests
requests==2.31.0
